server 파일(store.js) 분리 -> server 폴더
