+++ public/js/component/card.js
@@ -8,9 +8,9 @@ import { findColumnStatusByCard, findCardHeaderName } from "./column.js";
 import { menuLogAdd, menuLogUpdate, menuLogDeleteAll } from "./menu.js";
 import { turnOnModal } from "./modal.js";
 import { makeCardDragEvent } from "../drag/addDragEvent.js";
-import { addJSONData, deleteJSONData } from "../store/store.js"; 
+import { deleteJSONData } from "../../../server/DELETE.js";
 import { cardTemplate, newCardTemplate } from "../templates/template.js";
-import { idGenerator } from "../common/IDGenerator.js";
+import { addJSONData } from "../../../server/POST.js";
 
 let $chosenCard = "";
 let registering = false;
