+++ public/js/init.js
@@ -2,7 +2,7 @@ import { loadInitialData } from "./common/commonFunction.js";
 import { 
     addEventToMenuBtns, addEventToModalButtons, addEventToColumnAddButton, addEventToFabBtn 
 } from "./component/button.js";
-import { getAllJSONData } from "./store/store.js";
+import { getAllJSONData } from "../../server/GET.js";
 
 // 웹 페이지에 필요한 이벤트를 추가합니다.
 function addEventsToWebsite() {
